requirements.txt
# Core
langchain>=0.2
langchain-core>=1.2
langchain-classic>=0.1
langchain-text-splitters>=1.0
langchain-google-genai>=1.0
langchain-groq>=1.0
langchain-chroma>=1.0

# Vector DB
chromadb>=1.0

# API & Utilities
requests>=2.28
python-dotenv>=1.0

# Web UI
flask>=3.0
